[MISC] Move seqan3::views::chunk to seqan3/utility/views/
