+++ test/unit/core/algorithm/detail/execution_handler_template.hpp
@@ -15,9 +15,10 @@
 #include <gtest/gtest.h>
 
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
-#include <seqan3/range/views/chunk.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
 #include <seqan3/range/views/zip.hpp>
+#include <seqan3/utility/views/chunk.hpp>
+
 #include <seqan3/test/performance/sequence_generator.hpp>
 
 //Forward declaration
